requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.13
PyJWT==2.8.0 